refactor: split campaign instance state operations
